\name{dlmodeler-internal}
\alias{dlmodeler.bdiag}
\alias{dlmodeler.timevar.fun}
\alias{dlmodeler.filter.KFAS}
\alias{dlmodeler.filter.FKF}
\alias{dlmodeler.filter.dlm}
\alias{dlmodeler.smooth.KFAS}
\alias{dlmodeler.smooth.dlm}
\alias{dlmodeler.build.unknowns}
\alias{dlmodeler.build.function}
\alias{dlmodeler.countweekdays}
\alias{dlmodeler.build.constant}
\alias{dlmodeler.constant}
\title{
For internal use only
}
\description{
Functions for internal use only.
}
\keyword{ internal }
