\name{dlmodeler.yeardays}
\alias{dlmodeler.yeardays}
\title{
Return the number of days and weekdays in a given year
}
\description{
TODO
}
\usage{
dlmodeler.yeardays(year)
}
\arguments{
  \item{year}{year.}
}
\details{
TODO
}
\value{
TODO
}
\author{
Cyrille Szymanski <cnszym@gmail.com>
}
\note{
TODO
}

\seealso{
\code{\link{dlmodeler.build}}
}
\examples{
## TODO
}
\keyword{ dlm }
\keyword{ print }
